noCATfreq
  * change status type from int to long long
  * save and restore as a union of int and uint
  * allows maximum value of 9.999.999.999

+++ src/include/status.h
@@ -51,7 +51,7 @@ struct status {
 	int		rigW;
 	int		rigH;
 	int		carrier;
-	int		noCATfreq;
+	long long	noCATfreq;
 	std::string	noCATmode;
 	std::string noCATwidth;
 	int		mag;
